Add the Forms List page
After a user logs in the first thing they should be presented with is
the Forms page showing a list of all the forms they have permission to
interact with. At this time it will be any forms that they created.

This also involved us adding the JWT token generation for interaction
with the Service Token Cache via the Metadata API.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>