Specify remote docker version for CircleCI builds
If we do not specify the remote docker version CircleCI will default to
use Docker 17.0.9. This version does not play nice with node versions
14.9.0 or later. Consequently it throws a `EPERM: operation not permitted`
when trying to copy files.

Specifying a later Docker version aids in compatibility with node for
now allows the `yarn install` to work correctly.

Also add the `--production` flag to install.

Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>