Add edit page link in the page flow (form pages) page
I had to delete the layout metadata presenter because the layout
depends on one service running where on the editor we are running
multiple forms.

Also I had to leave the routes to be:

edit_page GET    /services/:id/pages/:page_url/edit(.:format)

because the service metadata is getting the service via
params[:id] so we should leave as params[:id] for nested routes.