Add acceptance tests for the editor
The base idea is to have tests where we can run locally or against
any environment (given a editor url).

* ENV vars *

Two env files are used:

1. one for running with local containers
2. other pointing to the test environment

*Local*

Added all necessary containers for the editor

- editor-db
- editor-app
- metadata-db
- metadata-app
- service-token-cache-app
- service-token-cache-redis

This will spin up the containers and run the tests.

* CI *

To run this you need to have the user and password setup on CircleCI
(I already added) or if you run on your local machine that points to
the test environment you can as well.
If you face 401 on the tests it is because the user env var and pass
aren't set.

*Caveats*

I have to make a specific dockerfile for acceptance test because we
need chrome driver and other packages to run.