Add brakeman ignore file for a couple of the warnings
This adds a brakeman ignore file with notes for 2 warnings that we
currently are ok ignoring.

The `permit!` we allow because the schemas are being used to control the
allowed properties that make it into the Metadata API.

The use of the URL param to find the next page is something that we will
be looking into changing shortly.