Add page uuid call to page creation
This fixes the test environment.

Page creation requires to redirect the user to edit the page that was
created. So it return the last page uuid.
This will be a problem when we need to create a page and insert in the
middle of the service but it solves the problem now.