Add acceptance tests
For the acceptance tests to identify which fields belong to the 'Single Question' form and which belonged to the 'Check Answers' form,
I have added an id in the forms.

Syntactically, I am unsure if this is the most elegant solution so please do let me know if there is a more elgant solution.