Add publish page
This is the part 1 of the publishing mechanism. This part involves
just adding the interface and preparing the cloud platform
setup in order for the editor to have permissions to provision
pods in the services namespace.

This adds the model (and the db table) without index just yet,
the controller and the empty job - without any job adapter yet.