Add publish mechanism
PublishServiceJob uses this classes:

  - PublishService to find the record in the database

  - ServiceProvisioner to use as a rich object to handle all the
    necessary configuration

  - Publisher to orchestrate the publishing mechanism by steps. The
    publisher will update the publish service status in the database
    while calls the adapter to handle the publish logic.

    - Adapters: The default adapter is CloudPlatform one that needs to
    respond to the following methods: pre_publishing, publishing,
    post_publishing and completed.

Details of the implementation

1. Each step uses the active support notification mechanism, which
the subscribers are added on initialize, so in the Rails logs we
can have the amount of time for each step:

Example:

Event 'publisher.pre_publishing' duration: (0.077 s)
Event 'publisher.publishing' duration: (0.077 s)
Event 'publisher.post_publishing' duration: (0.077 s)
Event 'publisher.completed' duration: (0.077 s)
Event 'publisher.main' duration: (0.077 s)

The 'publisher.main' is the amount of the whole process of publishing
and the others are the duration for each step.

2. I created an initializer platform_enviroments which contains
static data for each of our environments.

3. Created a RSpec matcher to test the generation of the config in
the tmp dir. The matcher verifies the existence and the content
of the file.

Next steps

Add data to generated config map
Add metadata to the environment variable (in the editor and the runner)
Generate the private and public key
Add readiness probe to the runner (/health url)
Add the publishing, post_publishing and completed step