Add private public key to services
After the service is created in the metadata API, the editor
will generate a private/public key pair and insert in the database
table called 'service_configurations'.

When the user clicks to publish, the job will look through
the service configurations searching by the service id and the
deployment environment (publish to test, publish to live) and then
pass to the service provisioner which is the rich object which
holds data.

The keys are stored in base64 format and added to the config map