Encrypt username and password for basic authentication
This adds the username and password for basic authentication as well as
encrypting all the service configuration values in the DB.

This includes the private and public keys. However the public key which
resides in the config still needs to be base64 encoded when written to
the config_map.yaml. This is because the datastore ServiceTokenCacheClient
does a base64 decode after retrieving the key. The service token cache
just looks for the key 'ENCODED_PUBLIC_KEY' in the config map and passes
back to the requesting app.

The 'ENCODED_' part of both keys relates to the time they are written to
the yaml files, _not_ when they are written to the DB.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>