Remove SENTRY_CURRENT_ENV
As a result of having 4 different namespace environments as opposed to
the regular 2 we structure our sentry projects differently.

We have 2 sentry projects per app, one for test and one for live.
The test one covers anything in the test-dev, test-production and
live-dev namespaces whereas live is exclusively for live-production
only. Therefore it's fine to let SENTRY_CURRENT_ENV use whatever
RAILS_ENV is set to as it does by default.