Adding hint text to sections
Added govuk-list class to the ul to have it aligned left.
Added hint copy.
Added way to many <br /> and awful inline style to get the spacing needed without messing up the css file as I'm not sure where it is…. and I figured that some ugly code in one place was easier to handle than mess across many files.
There was a bit of inconsistency between the variables naming for the first section (uses _) and the second (uses .), I continued it as didn't feel confident fixing it on some other pages.