Add user authentication to preview feature
The preview comes from the metadata presenter gem and has a superclass
inside of the editor app (PreviewController).

So by making the preview controller to inherit from permission
controller we make the user to be redirect to the login path

I had to chage the `root_path` because it makes the user to redirect to
root path in the in the context of the metadata
presenter gem (/services/:id/preview) which creates a infinite
redirection.

Adding the FbEditor::Application.url_helpers.root_path makes it return
'/' which is what we want.