Fix preview bug
When the user tries to preview the form the application rejects
when the user is going to submit an answer.

The reason is that Rails does two autenticity token verifications:
One for the app and one per action. So we disable the per action
verification because the preview feature and we still have
the autenticity token app verification.