Interact with username and password fields using JS
As a result of only having a single tenant for the moment in Auth0 we
need to be able to interact with the username and password field which
are hidden on the login screen.

The live editor should not show these fields as they will not be used by
an actual user, only by our acceptance tests. Unfortunately Auth0 does
not provide the facility to turn the fields on and off _if_ you have
chosen to use Universal Login, which we have.