Add standalone pages to new service generator
We need to have the ability to add standalone pages that are not part of
the flow. For now these are just the footer pages; cookies,
accessibility and privacy. They need to be there when a new service is
generated.

Instead of creating a standalone specific generator we adjusted the new
page generator to also be able to return just the page metadata instead
of returning the entire updated service metadata like it does when a
standard flow page is added.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>