Make the standalone pages editable
Configured the update and destroy methods in the MetadataUpdater to cater for standalone pages.
This PR allows the functionality to delete a standalone page, although this will not be surfaced in the UI so users will not be able to delete standalone pages.

Co-authored by: Brendan Butler <brendan.butler@digital.justice.gov.uk>