Use Node 14.16.1 LTS
We have some descrepencies with node versions in our containers but also
when developing locally. The README says use version 10 but the
container is installing 12 . . .

Lock the Node version to 14.16.1 as that is the current LTS version
which will be end of life in 2023.

Update the dockerfiles to install the specific node version and also use
a .npmrc file to lock the node version to the one we want.