Make extra_components to be editable
Before this commit the page[extra_components] were not setting a name
which makes the extra_components to be nil and send to the metadata
api which validates the schema and returns an error message about
the schema being invalid.

Adding the extra_components to behave in the same way of components
make the thing to be editable.