Make sure that component ID is unique across component collections
Since we now have both components and extra_components collections we
need to make sure that the component ID remains unique across both.

Therefore pass in all components into the new component generator which
will then increment the ID correctly.

At the same time add a uuid to the component object in case we need to
use it in the future.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>